+++ pkg/oci/mock.go
@@ -54,7 +54,3 @@ func (m *MockClient) GetManifest(ctx context.Context, dgst digest.Digest) ([]byt
 func (m *MockClient) GetBlob(ctx context.Context, dgst digest.Digest) (io.ReadCloser, error) {
 	return nil, nil
 }
-
-func (m *MockClient) CopyLayer(ctx context.Context, dgst digest.Digest, dst io.Writer) error {
-	return nil
-}