+++ pkg/oci/oci.go
@@ -21,6 +21,4 @@ type Client interface {
 	Size(ctx context.Context, dgst digest.Digest) (int64, error)
 	GetManifest(ctx context.Context, dgst digest.Digest) ([]byte, string, error)
 	GetBlob(ctx context.Context, dgst digest.Digest) (io.ReadCloser, error)
-	// Deprecated: Use GetBlob.
-	CopyLayer(ctx context.Context, dgst digest.Digest, dst io.Writer) error
 }